split task logic to multiple file
